Tweaks for twist axis
This is mostly just incorporating some of the lessons of improving
paddle support for stella. This means raising the damping
substantially, and applying a filter on all small deltas.

There are a few minor changes to sort out and define the signs of
angles, getting rid of the last weirdness there. The car's theta angle
increases anticlockwise, as is standard; the twist axis on the
joystick increases clockwise, which matches the natural sense of the
controller.